React refactoring some GETs changed to POSTs (complying with http standard) Dockerfile for react (compose is required)
