+++ worker-ui/public/manifest.json
@@ -1,6 +1,6 @@
 {
-  "short_name": "React App",
-  "name": "Create React App Sample",
+  "short_name": "KIBe",
+  "name": "KIBe Worker App - KIBe Alkalmazott App",
   "icons": [
     {
       "src": "favicon.ico",
